AAE-35733 Use controller advice for both controllers, fix tests
