+++ activiti-cloud-query-service/activiti-cloud-services-query/activiti-cloud-services-query-rest/src/test/java/org/activiti/cloud/services/query/rest/AbstractTaskControllerIT.java
@@ -24,6 +24,7 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 
 import java.math.BigDecimal;
 import java.util.Date;
@@ -79,6 +80,37 @@ public abstract class AbstractTaskControllerIT {
     @AfterEach
     public void cleanUp() {
         queryTestUtils.cleanUp();
+    }
+
+    @Test
+    void should_return400_whenInvalidSearchParameterIsProvided() {
+        String missingSortField =
+            """
+        {
+            "sort": {
+                "direction": "ASC",
+                "isProcessVariable": false,
+                "processDefinitionKey": null,
+                "type": "bigdecimal",
+                "processVariable": false
+            }
+        }""";
+
+        given()
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(missingSortField)
+            .when()
+            .post(getSearchEndpointHttpPost())
+            .then()
+            .statusCode(400)
+            .expect(
+                content()
+                    .string(
+                        is(
+                            "Invalid search parameter: Could not resolve attribute 'null' of 'org.activiti.cloud.services.query.model.TaskEntity'"
+                        )
+                    )
+            );
     }
 
     @Test
